+++ changelog.md
@@ -16,6 +16,7 @@ Codion Change Log
 - InputDialogBuilder.inputValid() and inputValidator() renamed valid() and validator() respectively.
 ### is.codion.swing.fraework.ui
 - EntitySearchFieldPanel.Builder.includeAddButton() and includeEditButton() bug fixed, exception now thrown as per javadoc.
+- EntityEditPanel.Config.editPanel() added.
 ### is.codion.swing.framework.server.monitor
 - ConnectionPoolMonitor, pool cleanup interval now milliseconds based.